requirements.txt
APScheduler>=3.8.1,<3.9.0
cx-Oracle>=8.2.1,<8.3.0
requests>=2.26.0,<2.27.0
pandas>=1.2.4,<1.3.0
psycopg2-binary>=2.9.1,<2.10.0
odfpy>=1.4.1,<1.5.0